refactor: split checkouts into two files
